Moving GetKustoQueryUriAsync to IKustoClient and making it async
