+++ src/Diagnostics.DataProviders/Interfaces/IKustoClient.cs
@@ -11,5 +11,6 @@ namespace Diagnostics.DataProviders
     public interface IKustoClient
     {
         Task<DataTable> ExecuteQueryAsync(string query, string stampName, string requestId = null, string operationName = null);
+        Task<string> GetKustoQueryUriAsync(string stampName, string query);
     }
 }
